requirements.txt
python-telegram-bot[ext]==21.*
motor==3.*
pymongo==4.*
python-dotenv==1.*
